backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
openai==1.35.3
youtube-transcript-api==1.2.4
tiktoken==0.7.0
httpx==0.27.0
python-dotenv==1.0.1
httpx==0.27.0
